pyproject.toml
[project]
name = "limerick-benchmark"
version = "0.1.0"
description = "Benchmark suite for evaluating LLM coding ability on local Apple Silicon"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "rich>=13.0",
    "litellm>=1.70",
    "psutil>=6.0",
    "aiohttp>=3.10",
]
